Fixing bug in SDL2 branch and warnings fixes.
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_softfloat@1848 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/snd.c
@@ -268,12 +268,12 @@ static void snd_save_volume_reg(void) {
         vol_data = SND_MAX_VOL;
     }
     if (chan_lr & 1) {
-        Log_Printf(LOG_WARN, "[Sound] Setting gain of left channel to %d dB", vol_data * -2);
+        Log_Printf(LOG_WARN, "[Sound] Setting gain of left channel to %d dB", (int)vol_data * -2);
         sndout_state.attenuation[0] = vol_data;
         sndout_state.volume[0] = snd_get_volume_factor(vol_data);
     }
     if (chan_lr & 2) {
-        Log_Printf(LOG_WARN, "[Sound] Setting gain of right channel to %d dB", vol_data * -2);
+        Log_Printf(LOG_WARN, "[Sound] Setting gain of right channel to %d dB", (int)vol_data * -2);
         sndout_state.attenuation[1] = vol_data;
         sndout_state.volume[1] = snd_get_volume_factor(vol_data);
     }
